refactorizado controllador de ChallengeMicroservice
