+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/LanguageController.java
@@ -1,0 +1,37 @@
+package com.itachallenge.challenge.controller;
+
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.LanguageDto;
+import com.itachallenge.challenge.service.IChallengeService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
+@RestController
+@Validated
+@RequiredArgsConstructor
+@RequestMapping(value = "/itachallenge/api/v1/language")
+public class LanguageController {
+
+    private IChallengeService challengeService;
+
+    @GetMapping("/language")
+    @Operation(
+            operationId = "Get all the stored languages into the Database.",
+            summary = "Get to see all id language and name.",
+            description = "Requesting all the languages through the URI from the database.",
+            responses = {
+                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = GenericResultDto.class), mediaType = "application/json")}),
+            }
+    )
+    public Mono<GenericResultDto<LanguageDto>> getAllLanguages() {
+        return challengeService.getAllLanguages();
+    }
+}
