+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/controller/TagController.java
@@ -1,0 +1,37 @@
+package com.itachallenge.challenge.controller;
+
+import com.itachallenge.challenge.dto.GenericResultDto;
+import com.itachallenge.challenge.dto.TagDto;
+import com.itachallenge.challenge.service.ITagService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import reactor.core.publisher.Mono;
+
+@RestController
+@Validated
+@RequiredArgsConstructor
+@RequestMapping(value = "/itachallenge/api/v1/tag")
+public class TagController {
+
+    private ITagService tagService;
+
+    @GetMapping("/tags")
+    @Operation(
+            operationId = "Get all stored tags from the Database for FrontEnd can print them.",
+            summary = "Get to see all id tags, name and description.",
+            description = "Requesting all the tags through the URI from the database.",
+            responses = {
+                    @ApiResponse(responseCode = "200", content = {@Content(schema = @Schema(implementation = GenericResultDto.class), mediaType = "application/json")}),
+            }
+    )
+    public Mono<GenericResultDto<TagDto>> getAllTags() {
+        return tagService.getAllTags();
+    }
+}
